factor out scope string manipulation
