+++ src/features/profile/APITokens/APITokenDrawer.tsx
@@ -24,6 +24,14 @@ import Select from 'src/components/Select';
 import { dateFormat } from 'src/events';
 import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
 
+import {
+  Permission,
+  scopeStringToPermTuples,
+  permTuplesToScopeString,
+} from './utils';
+
+type Expiry = [string, string];
+
 type ClassNames = 'permsTable';
 
 const styles: StyleRulesCallback<ClassNames> = (theme: Theme) => ({
@@ -33,9 +41,6 @@ const styles: StyleRulesCallback<ClassNames> = (theme: Theme) => ({
 });
 
 export type DrawerMode = 'view' | 'edit' | 'create';
-
-type Permission = [string, number];
-type Expiry = [string, string];
 
 interface RadioButton extends HTMLInputElement {
   name: string;
@@ -92,148 +97,14 @@ class APITokenDrawer extends React.Component<CombinedProps, State> {
     }
   }
 
-  static perms = [
-    'account',
-    'domains',
-    'events',
-    'images',
-    'ips',
-    'linodes',
-    'longview',
-    'nodebalancers',
-    'stackscripts',
-    'volumes',
-  ];
-
-  static levelMap = {
-    none: 0,
-    read_only: 1,
-    read_write: 2,
-    view: 1,
-    modify: 2,
-    create: 2,
-    delete: 2,
-  };
-
-  static inverseLevelMap = [
-    'none',
-    'read_only',
-    'read_write',
-  ];
-
-  /**
-   * This function accepts scopes strings directly from the API, which have the following format:
-   * "linodes:delete,domains:modify,nodebalancers:modify,images:create,events:view,clients:view"
-   *
-   * It returns an array of 2-tuples in alphabetical order by scope name.
-   *
-   * Each 2-tuple has the format [<scopename>, <number>], where <number> is the permission level
-   * of the scope. These are the permission levels in order:
-   *
-   * None: 0
-   * ReadOnly: 1
-   * ReadWrite: 2
-   *
-   * These are old permission levels which must be mapped to the new levels
-   * None: 0
-   * View: 1
-   * Create: 2
-   * Modify: 2
-   * Delete: 2
-   *
-   * Each permission level gives a user access to all lower permission levels.
-   */
-  scopesToPermTuples(scopes: string): Permission[] {
-    if (scopes === '*') {
-      return APITokenDrawer.perms.map(perm => [perm, 2] as Permission);
-    }
-
-    const scopeMap = scopes.split(',').reduce(
-      (map, scopeStr) => {
-        const scopeTuple = scopeStr.split(':');
-        scopeTuple[1] = APITokenDrawer.levelMap[scopeTuple[1]];
-        map[scopeTuple[0]] = scopeTuple[1];
-        return map;
-      },
-      {},
-    );
-
-    const equivalentPerms = {
-      account: [
-        'tokens',
-        'clients',
-        'users',
-        'tickets',
-        'managed',
-      ],
-    };
-
-    /* TODO: Remove this logic once the API starts doing the deprecation mapping for us */
-    const combinedScopeMap = Object.keys(equivalentPerms).reduce(
-      (map: { [perm: string]: number }, perm: string) => {
-        const maxLevel = equivalentPerms[perm].reduce(
-          (level: number, eqPerm: string) => {
-            return Math.max(level, map[eqPerm] || APITokenDrawer.levelMap['none']);
-          },
-          map[perm] || APITokenDrawer.levelMap['none'],
-        );
-        map[perm] = maxLevel;
-        return map;
-      },
-      scopeMap,
-    );
-
-    const permTuples = APITokenDrawer.perms.reduce(
-      (tups: Permission[], permName: string): Permission[] => {
-        const tup = [
-          permName,
-          combinedScopeMap[permName] || APITokenDrawer.levelMap['none'],
-        ] as Permission;
-        return [...tups, tup];
-      },
-      [],
-    );
-
-    return permTuples;
-  }
-
-  allMaxPerm = (scopeTups: Permission[]) : boolean => {
-    if (scopeTups.length !== APITokenDrawer.perms.length) {
-      return false;
-    }
-    return scopeTups.reduce(
-      (acc: boolean, scopeTup: Permission) => {
-        return (scopeTup[1] === APITokenDrawer.levelMap.read_write) && acc;
-      },
-      true,
-    );
-  }
-
-  permTuplesToScopeString = (scopeTups: Permission[]): string => {
-    if (this.allMaxPerm(scopeTups)) {
-      return '*';
-    }
-    const joinedTups = scopeTups.reduce(
-      (acc, tup) => {
-        const level = APITokenDrawer.inverseLevelMap[tup[1]];
-        if (level !== 'none') {
-          return [...acc, [tup[0], level].join(':')];
-        }
-        return [...acc];
-      },
-      [],
-    );
-    return joinedTups.join(',');
-  }
-
   handleScopeChange = (e: React.SyntheticEvent<RadioButton>): void => {
-    const scopeTups = this.scopesToPermTuples(this.state.scopes);
+    const scopeTups = scopeStringToPermTuples(this.state.scopes);
     const targetIndex = scopeTups.findIndex(
       (scopeTup: Permission) => scopeTup[0] === e.currentTarget.name);
     if (targetIndex !== undefined) {
       scopeTups[targetIndex][1] = +(e.currentTarget.value);
     }
-    this.setState({ scopes: this.permTuplesToScopeString(scopeTups) });
+    this.setState({ scopes: permTuplesToScopeString(scopeTups) });
   }
 
   permNameMap = {
@@ -264,7 +135,7 @@ class APITokenDrawer extends React.Component<CombinedProps, State> {
           </TableRow>
         </TableHead>
         <TableBody>
-          {this.scopesToPermTuples(scopes).map(
+          {scopeStringToPermTuples(scopes).map(
             (scopeTup) => {
               return (
                 <TableRow key={scopeTup[0]}>
